requirements.txt
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-polardb==1.8.10
allure-cmd==0.0.5
allure-pytest==2.9.43
allure-python-commons==2.9.43
Appium-Python-Client==1.2.0
APScheduler==3.8.1
atomicwrites==1.4.0
attrs==21.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
certifi==2018.4.16
cffi==1.14.6
chardet==3.0.4
click==6.7
colorama==0.3.9
crypto==1.4.1
cryptography==3.4.7
cycler==0.11.0
dataclasses==0.6
ddt==1.4.2
decorator==5.0.9
dominate==2.3.1
dukpy==0.2.3
enum34==1.1.10
et-xmlfile==1.1.0
execnet==1.9.0
fire==0.4.0
Flask==1.0.2
Flask-APScheduler==1.12.2
Flask-Bootstrap==3.3.7.1
fonttools==4.29.1
future==0.16.0
gitdb==4.0.7
GitPython==3.1.18
greenlet==1.1.2
httplib2==0.20.2
humanize==3.12.0
idna==2.7
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==0.24
javascripthon==0.11
Jinja2==2.11.3
jmespath==0.10.0
jsonpath==0.82
jsonpath-rw==1.4.0
junit-report==0.1.24
junit-xml==1.9
jupyter-echarts-pypkg==0.1.1
kiwisolver==1.3.2
lml==0.0.2
logzero==1.7.0
lxml==4.6.3
macropy3==1.1.0b2
MarkupSafe==1.0
matplotlib==3.5.1
more-itertools==8.8.0
mysql==0.0.2
mysql-connector-python==8.0.12
mysqlclient==2.0.3
Naked==0.1.31
namedlist==1.8
numpy==1.20.2
openpyxl==3.0.7
packaging==21.0
pandas==1.2.4
paramiko==2.7.2
Pillow==8.3.2
pluggy==0.13.1
ply==3.11
prettytable==2.4.0
progress==1.6
protobuf==3.6.0
py==1.10.0
pycparser==2.20
pycryptodome==3.12.0
pydash==5.0.0
pyDes==2.0.1
pyecharts==0.5.5
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pymongo==4.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==3.7.1
pytest-assume==2.4.3
pytest-fixture-config==1.7.0
pytest-fixtures==0.1.0
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-rerun==0.0.1
pytest-rerunfailures==10.1
pytest-testreport==1.0.9
pytest-xdist==2.3.0
PyTestReport==0.2.1
python-dateutil==2.8.1
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
requests==2.24.0
retry==0.9.2
retrying==1.3.3
scipy==1.8.0
selenium==3.141.0
shellescape==3.8.1
simplejson==3.17.5
six==1.11.0
smmap==4.0.0
SQLAlchemy==1.4.27
sshtunnel==0.4.0
str2bool==1.1
string-to-bool==0.0.2
sweetviz==2.1.3
termcolor==1.1.0
toml==0.10.2
tqdm==4.62.3
trafaret==2.1.0
tzdata==2021.5
tzlocal==4.1
uiautomator2==0.0.3
unittest-asserts==0.1.0
unittestreport==1.3.2
urllib3==1.25.11
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.0.2
zipp==3.7.0
